fixed api and bugs
