+++ bundesliga/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.1.2'
+__version__ = '0.2'
