+++ bundesliga/openligadb.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
 
 import json
-import urllib2
-from datetime import datetime
-from helpers import *
+import re
+from urllib.request import urlopen
 
 
 class OpenLigaDB:
@@ -16,26 +15,28 @@ class OpenLigaDB:
     FUSSBALL_SPORT_ID = 1
 
     def __init__(self):
-        self.openLigaDBApiUrl = 'http://openligadb-json-api.herokuapp.com/api'
+        # self.openLigaDBApiUrl = 'http://openligadb-json-api.herokuapp.com/api'
+        self.openLigaDBApiUrl = 'https://www.openligadb.de/api'
 
     def getMatchdayResults(self, matchday=0, season="", league=""):
         """
         Returns the results for the given matchday
         """
         if matchday == 0:
-            matchday = self.getNextMatchday();
+            matchday = self.getNextMatchday()
 
         if season == "":
-            saison = self.getSeason()
+            season = self.getSeason()
 
         if league == "":
             league = self.ERSTE_LIGA
 
-        requestUrl = self.openLigaDBApiUrl + '/matchdata_by_group_league_saison'
-        requestUrl += '?group_order_id=%s&league_saison=%s&league_shortcut=%s' % (matchday, season, league)
-        data = json.load(urllib2.urlopen(requestUrl))
+        requestUrl = (self.openLigaDBApiUrl + '/getmatchdata/' +
+                      league + '/' + season + '/' + str(matchday))
+        data = json.load(urlopen(requestUrl))
 
-        return data['GetMatchdataByGroupLeagueSaisonResult']
+        # return data['GetMatchdataByGroupLeagueSaisonResult']
+        return data
 
     def getTable(self, season, league=ERSTE_LIGA):
         """
@@ -45,8 +46,8 @@ class OpenLigaDB:
         teams = self.getTeams(season, league)
         table = {}
         for team in teams:
-            table[team['teamID']] = {
-                    'team_name': team['teamName'],
+            table[team['TeamId']] = {
+                    'team_name': team['TeamName'],
                     'points': 0,
                     'wins': 0,
                     'losses': 0,
@@ -58,13 +59,14 @@ class OpenLigaDB:
         matchData = self.getMatchDataByLeagueSaison(season, league)
 
         for match in matchData:
-            if not match['matchIsFinished']:
+            if not match['MatchIsFinished']:
                 continue
 
-            team1 = match['idTeam1']
-            team2 = match['idTeam2']
-            goals_team1 = int(match['pointsTeam1'])
-            goals_team2 = int(match['pointsTeam2'])
+            team1 = match['Team1']['TeamId']
+            team2 = match['Team2']['TeamId']
+            # team2 = match['idTeam2']
+            goals_team1 = int(match['MatchResults'][-1]['PointsTeam1'])
+            goals_team2 = int(match['MatchResults'][-1]['PointsTeam2'])
 
             teamData1 = table[team1]
             teamData2 = table[team2]
@@ -88,16 +90,22 @@ class OpenLigaDB:
                 teamData1['draws'] += 1
                 teamData2['draws'] += 1
 
-        return sorted([value for key, value in table.iteritems()], key=lambda k: k['points'], reverse=True)
+        return sorted([value for key, value in table.items()],
+                      key=lambda k: (k['points'],
+                                     k['goals']-k['received_goals'],
+                                     k['goals']),
+                      reverse=True)
 
-    def getNextMatchday(self):
+    def getNextMatchday(self, league=ERSTE_LIGA):
         """
         Returns the next matchday (consider that the next matchday could be
         the current matchday).
         """
-        requestUrl = self.openLigaDBApiUrl + '/current_group?league_shortcut=bl1'
-        data = json.load(urllib2.urlopen(requestUrl))
-        return data['GetCurrentGroupResult']['groupOrderID']
+        requestUrl = (self.openLigaDBApiUrl +
+                      '/getmatchdata/' + league)
+        data = json.load(urlopen(requestUrl))
+        # return data['GetCurrentGroupResult']['groupOrderID']
+        return data[0]['Group']['GroupOrderID']
 
     def getRecentMatchday(self):
         """
@@ -108,67 +116,56 @@ class OpenLigaDB:
             previousMatchday = 1
         return str(previousMatchday)
 
-    def getSeasons(self, league='bl1'):
-        """
-        Returns the current saison.
-        """
-        requestUrl = self.openLigaDBApiUrl + '/seasons?league=' + league
-        data = json.load(urllib2.urlopen(requestUrl))
-        return data
+    def getCurrentSeason(self, league=ERSTE_LIGA):
+        requestUrl = (self.openLigaDBApiUrl +
+                      '/getmatchdata/' + league)
+        data = json.load(urlopen(requestUrl))
+        leagueName = data[0]['LeagueName']
+        season = re.sub(r'.*?(\d{4}).*', r'\1', leagueName)
 
-    def getCurrentSeason(self, league='bl1'):
-        seasons = self.getSeasons(league)
-        now = datetime.datetime.now()
-
-        for key in seasons:
-            season = seasons[key]
-            start = from_utc(season['start'])
-            end = from_utc(season['end'])
-
-            if now > start and now <= end:
-                return season
-
-        return getYear(now)
-
+        return season
 
     def getTeams(self, season, league=ERSTE_LIGA):
         """
         Returns a list of all teams for the given season and league.
         """
-        requestUrl = self.openLigaDBApiUrl + '/teams_by_league_saison?league_saison=%s&league_shortcut=%s' % (season, league)
-        data = json.load(urllib2.urlopen(requestUrl))
+        requestUrl = (self.openLigaDBApiUrl +
+                      '/getavailableteams/' + league + '/' + season)
+        data = json.load(urlopen(requestUrl))
 
         teams = []
-        for team in data['GetTeamsByLeagueSaisonResult']:
-            teams.append(team['Team'])
+        for team in data:
+            teams.append(team)
 
         return teams
 
     def getMatchDataByLeagueSaison(self, season, league=ERSTE_LIGA):
         """
         """
-        requestUrl = self.openLigaDBApiUrl + '/matchdata_by_league_saison?league_saison=%s&league_shortcut=%s' % (season, league)
-        data = json.load(urllib2.urlopen(requestUrl))
+        requestUrl = (self.openLigaDBApiUrl +
+                      '/getmatchdata/' + league + '/' + season)
+        data = json.load(urlopen(requestUrl))
 
         matchData = []
-        for match in data['GetMatchdataByLeagueSaisonResult']:
-            matchData.append(match['Matchdata'])
+        for match in data:
+            matchData.append(match)
 
         return matchData
 
     def getMatchesByTeam(self, team):
         pass
 
-    def getAvailLeagues(self):
-        requestUrl = self.openLigaDBApiUrl + '/avail_leagues'
-        data = json.load(urllib2.urlopen(requestUrl))
+    # depreceated
+    # def getAvailLeagues(self):
+    #     requestUrl = self.openLigaDBApiUrl + '/avail_leagues'
+    #     data = json.load(urlopen(requestUrl))
 
-        leagues = []
-        for league in data['GetAvailLeaguesResult']:
-            league = league['League']
-            if (league['leagueSportID'] == self.FUSSBALL_SPORT_ID
-                and 'test' not in league['leagueName'].lower()
-                and league['leagueID'] != 532):
-                leagues.append(league)
+    #     leagues = []
+    #     for league in data['GetAvailLeaguesResult']:
+    #         league = league['League']
+    #         if (league['leagueSportID'] == self.FUSSBALL_SPORT_ID and
+    #                 'test' not in league['leagueName'].lower() and
+    #                 league['leagueID'] != 532):
+    #             leagues.append(league)
 
-        return leagues
+    #     return leagues
